backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
mysql-connector-python==8.4.0
pydantic==2.9.2
pydantic-settings==2.5.2
ollama==0.3.3
python-dotenv==1.0.1
